+++ sound-mind-studio/tests/test_main_window.h
@@ -178,4 +178,5 @@ private slots:
     void settingANewProjectResetsPathModeToOff();
     void pastedContentIsPickableAndMovable();
     void modifyingAPaintedStrokeAfterCuttingOverItKeepsTheCutRegionSilenced();
+    void bringPickedObjectToFrontMovesItAboveLaterStrokesOnTheSameLayer();
 };